requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.2
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imutils==0.5.4
ipykernel==6.22.0
ipython==8.11.0
itsdangerous==2.1.2
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
pywin32==305
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.14.1
